requirements.txt
msgraph-beta-sdk>=1.0.0
azure-identity>=1.15.0
apscheduler>=3.10.0
